+++ zcash_client_backend/src/data_api/testing.rs
@@ -2946,7 +2946,7 @@ impl WalletWrite for MockWalletDb {
     /// Adds a transparent UTXO received by the wallet to the data store.
     fn put_received_transparent_utxo(
         &mut self,
-        _output: &WalletTransparentOutput,
+        _output: &WalletTransparentOutput<Self::AccountId>,
     ) -> Result<Self::UtxoRef, Self::Error> {
         Ok(0)
     }