+++ zcash_client_backend/src/data_api/testing/pool.rs
@@ -3339,6 +3339,8 @@ where
         spent_outpoint.clone(),
         TxOut::new(Zatoshis::const_from_u64(100000), taddr.script().into()),
         Some(h),
+        crate::TransferType::Incoming,
+        account.id(),
         Some(TransparentKeyScope::EXTERNAL),
     )
     .unwrap();
@@ -5176,6 +5178,8 @@ pub fn wallet_recovery_computes_fees<T: ShieldedPoolTester, DsF: DataStoreFactor
             outpoint,
             t_bundle.vout[0].clone(),
             Some(h),
+            crate::TransferType::Incoming,
+            dest_account_id,
             Some(TransparentKeyScope::EXTERNAL),
         )
         .unwrap();