+++ zcash_client_backend/src/data_api/testing/transparent.rs
@@ -125,6 +125,8 @@ where
         outpoint.clone(),
         txout.clone(),
         Some(height_1),
+        crate::TransferType::Incoming,
+        account_id,
         Some(TransparentKeyScope::EXTERNAL),
     )
     .unwrap();
@@ -156,6 +158,8 @@ where
         outpoint,
         txout,
         Some(height_2),
+        crate::TransferType::Incoming,
+        account_id,
         Some(TransparentKeyScope::EXTERNAL),
     )
     .unwrap();
@@ -242,6 +246,8 @@ where
         OutPoint::fake(),
         txout,
         Some(height),
+        crate::TransferType::Incoming,
+        account.id(),
         Some(TransparentKeyScope::EXTERNAL),
     )
     .unwrap();
@@ -389,6 +395,8 @@ where
         OutPoint::fake(),
         txout,
         Some(height),
+        crate::TransferType::Incoming,
+        account.id(),
         Some(TransparentKeyScope::EXTERNAL),
     )
     .unwrap();
@@ -757,7 +765,15 @@ where
     let value = Zatoshis::const_from_u64(50_000);
     let outpoint = OutPoint::fake();
     let txout = TxOut::new(value, taddr.script().into());
-    let utxo = WalletTransparentOutput::from_parts(outpoint, txout, Some(height), None).unwrap();
+    let utxo = WalletTransparentOutput::from_parts(
+        outpoint,
+        txout,
+        Some(height),
+        crate::TransferType::Incoming,
+        account_id,
+        None,
+    )
+    .unwrap();
     st.wallet_mut()
         .put_received_transparent_utxo(&utxo)
         .unwrap();
@@ -828,8 +844,15 @@ where
     let value = Zatoshis::from_u64(100000).unwrap();
     let height = st.wallet().chain_height().unwrap().unwrap();
     let txout = TxOut::new(value, taddr.script().into());
-    let utxo =
-        WalletTransparentOutput::from_parts(OutPoint::fake(), txout, Some(height), None).unwrap();
+    let utxo = WalletTransparentOutput::from_parts(
+        OutPoint::fake(),
+        txout,
+        Some(height),
+        crate::TransferType::Incoming,
+        account_id,
+        None,
+    )
+    .unwrap();
     st.wallet_mut()
         .put_received_transparent_utxo(&utxo)
         .unwrap();
@@ -1083,7 +1106,15 @@ where
     let value = Zatoshis::const_from_u64(50_000);
     let outpoint = OutPoint::fake();
     let txout = TxOut::new(value, taddr.script().into());
-    let utxo = WalletTransparentOutput::from_parts(outpoint, txout, Some(height), None).unwrap();
+    let utxo = WalletTransparentOutput::from_parts(
+        outpoint,
+        txout,
+        Some(height),
+        crate::TransferType::Incoming,
+        account_id,
+        None,
+    )
+    .unwrap();
     st.wallet_mut()
         .put_received_transparent_utxo(&utxo)
         .unwrap();
@@ -1151,8 +1182,15 @@ where
     let value = Zatoshis::from_u64(100000).unwrap();
     let height = st.wallet().chain_height().unwrap().unwrap();
     let txout = TxOut::new(value, taddr.script().into());
-    let utxo =
-        WalletTransparentOutput::from_parts(OutPoint::fake(), txout, Some(height), None).unwrap();
+    let utxo = WalletTransparentOutput::from_parts(
+        OutPoint::fake(),
+        txout,
+        Some(height),
+        crate::TransferType::Incoming,
+        account_id,
+        None,
+    )
+    .unwrap();
     st.wallet_mut()
         .put_received_transparent_utxo(&utxo)
         .unwrap();