+++ zcash_client_backend/src/proto.rs
@@ -750,7 +750,8 @@ impl proposal::Proposal {
                                                             PoolType::TRANSPARENT,
                                                             out.index,
                                                         )
-                                                    })?,
+                                                    })?
+                                                    .without_account_id(),
                                             );
                                         }
                                     }